The SVGicon is now generated using a path instead of an image file. This allows other props such as fillColor (which is based on theme) to be passed to the icon for the marker. Also reorganised the code directories
