+++ src/js/AppBar.js
@@ -3,7 +3,7 @@ import AppBar from 'material-ui/AppBar';
 import Drawer from './Drawer';
 import IconButton from 'material-ui/IconButton';
 import * as Str_en from './Strings_en';
-import BendigoLogo from './logo_nav.png';
+import BendigoLogo from '../images/logo_nav.png';
 
 
 export default class MainAppBar extends React.Component {
