+++ src/js/Drawer.js
@@ -7,7 +7,7 @@ import Divider from 'material-ui/Divider';
 
 import MapLayerIcon from 'material-ui/svg-icons/maps/layers';
 import MapLayerErrorIcon from 'material-ui/svg-icons/alert/error-outline';
-import logo from './logo.png';
+import logo from '../images/logo.png';
 import DialogMenuItem from './MenuDialogBox';
 import * as Str_en      from './Strings_en';
 
@@ -52,9 +52,11 @@ export default class DrawerUndocked extends React.Component{
         this.setState({open: !this.state.open})
         console.log("open:" + this.state);
     }
+
     handleDataSwap(endpoint, name){
         this.props.changeEndpoint(endpoint, name);
     }
+
     handleItemClick(name, endpoint){
         this.handleToggling()
         this.handleDataSwap(name, endpoint);
