+++ src/js/FlatLinkButton.js
